数据从int改成size_t，可以放指针来自定义数据类型 格式化代码

+++ linkedList.h
@@ -1,5 +1,5 @@
 typedef struct _ListNode {
-    int data;
+    size_t data;
     struct _ListNode* prevNode;
     struct _ListNode* nextNode;
 } *lNode;
@@ -14,13 +14,13 @@ typedef struct _List {
 list list_new();
 
 //Get data from list
-int list_get(list L, int pos);
+size_t list_get(list L, int pos);
 
 //Add node to list
-void list_push(list L, int pos, int data);
+void list_push(list L, int pos, size_t data);
 
 //Remove node from list
-int list_pop(list L, int pos);
+size_t list_pop(list L, int pos);
 
 //Free a list
 void list_free(list L);
